requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.1
filelock==3.0.12
gunicorn==20.0.4
Markdown==3.2.2
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
whitenoise==5.2.0
